backend/api/tests/conftest.py
import pytest
from django.contrib.auth.models import User
from rest_framework.test import APIClient

@pytest.fixture
def api_client():
    return APIClient()

@pytest.fixture
def user():
    return User.objects.create_user(username='john', password='password')

@pytest.fixture
def user2():
    return User.objects.create_user(username='jane', password='password2')

@pytest.fixture
def auth_client(api_client, user):
    api_client.force_authenticate(user=user)
    return api_client
